refactor(velvet): consolidate cross-panel input dispatch into one file
FiberCrossPanelEventDispatcher (synthetic bubbling OUT of a host panel) and
FiberCrossPanelPointerRouter (picking-order arbitration INTO a host panel) are the two ends
of the same pipe -- the same cross-panel input mechanism split across two files meant reading
either half in isolation missed how a cross-panel event actually travels, and scattered
related logic across files is its own bug risk (a change to one side's assumptions easy to
miss updating on the other). Merged into FiberCrossPanelInput.cs with no behavior change --
same two classes, same members, same call sites.

Full EditMode (2763/2763) and PlayMode (46/46) suites passed.
