--- Packages/com.velvet.core/Runtime/Reconciler/FiberCrossPanelEventDispatcher.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8345b6ca6f1b7474ca29487c2e974f96
