--- Packages/com.velvet.core/Runtime/Reconciler/FiberCrossPanelPointerRouter.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 990b85baef762426a860749cd11a668c
